[connector] fix: Remove undeterminism from notion connector signals
